Converted logging script to py3 and changed from pyawi to pyOWM

+++ BMP085.py
@@ -55,7 +55,7 @@ class Bmp085(object):
         X1 = ((UT - self.AC6) * self.AC5) >> 15
         X2 = (self.MC << 11) / (X1 + self.MD)
         B5 = X1 + X2
-        temp = ((B5 +8) >> 4) / 10.0
+        temp = (int(B5 + 8) >> 4) / 10.0
         # Ultra precise mesur mode for pressure.
         mode = 3
         # Reading raw pressure.
@@ -70,12 +70,12 @@ class Bmp085(object):
         UP = ((msb << 16) + (lsb << 8) + xlsb) >> (8 - mode)
         #print "UP", UP
         B6 = B5 - 4000
-        X1 = (self.B2 * ( B6 * B6 ) >> 12 ) >> 11
-        X2 = (self.AC2 * B6) >> 11
+        X1 = int(self.B2 * int( B6 * B6 ) >> 12 ) >> 11
+        X2 = int(self.AC2 * B6) >> 11
         X3 = X1 + X2
         B3 = (((self.AC1 * 4 + X3) << mode) + 2) / 4
-        X1 = (self.AC3 * B6) >> 13
-        X2 = (self.B1 * ((B6 * B6) >> 12)) >> 16
+        X1 = int(self.AC3 * B6) >> 13
+        X2 = int(self.B1 * (int(B6 * B6) >> 12)) >> 16
         X3 = ((X1 + X2) + 2) >> 2
         B4 = (self.AC4 * (X3 + 32768)) >> 15
         B7 = (UP - B3) * (50000 >> mode)
@@ -83,8 +83,8 @@ class Bmp085(object):
             p = (B7 * 2) / B4
         else:
             p = (B7 / B4) * 2
-        X1 = (p >> 8) * (p >> 8)
+        X1 = (int(p) >> 8) * (int(p) >> 8)
         X1 = (X1 * 3038) >> 16
-        X2 = (-7357 * p) >> 16
+        X2 = int(-7357 * p) >> 16
         p = p + ((X1 + X2 + 3791) >> 4)
         return (temp, p)
